refactor: consolidate examples into facade crate, remove docs from repo
Move 5 examples from sub-crates (kithara-audio, kithara-file, kithara-hls)
into the kithara facade crate with imports rewritten to use kithara::prelude::*.
Delete obsolete decode research utilities. Add AbrMode/AbrOptions to prelude.
Rename resource_play → resource_rodio. Remove docs/ from git, keep locally.
