+++ crates/kithara-decode/Cargo.toml
@@ -32,11 +32,3 @@ test-utils = ["unimock"]
 [dev-dependencies]
 rstest = { workspace = true }
 unimock = { workspace = true }
-
-[[example]]
-name = "prime_info_check"
-required-features = ["apple"]
-
-[[example]]
-name = "abr_switch_simulator_apple"
-required-features = ["apple"]
